+++ lib/gamefic-standard/entities/thing.rb
@@ -2,5 +2,4 @@
 
 Thing = Gamefic::Entity
 Thing.set_default itemized: true, portable: false
-Thing.include Gamefic::Standard::Grammar::Attributes
 Thing.include Gamefic::Standard::Standardized