+++ lib/gamefic-standard/grammar.rb
@@ -1,2 +1,0 @@
-require 'gamefic-standard/grammar/attributes'
-require 'gamefic-standard/grammar/pronoun'