+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add gamefic-what
 - Standard::Actions namespace
 - Room#connect accepts additional portal attributes
+- Add gamefic-conjugation
 
 ## 3.2.4 - September 10, 2024
 - Include Articles in Standard